chore: update test snapshots
Most of the updates are the result of the reimplementation of how we calculate the text edits.
Previously, the text edit is acquired from a naive text diff of the code before and after commit.
We now populate the text edit struct manually when applying changes.

Some of the updates are the reuslt of discarding leading trivia when removing a node.