feat(lint): add `ignoreNull` option to `noDoubleEquals` rule
Biome's `noDoubleEquals` rule initially made an exception when checking
against `null`.

This not something that is always wanted by linter users, some prefering
no exception to this rule in the name of explicitness.

The original `eslint` rule, `eqeqeq` has an `always` option to disable
specific handling of `null`. Here, I propose an `ignoreNull` option
instead, by default set to `true` to not break API (and as it may be
your recommended setting), because I find it much more explicit on what
it does than the more vague `always`.