feat: support OpenBSD
Use `std::alloc::System` for OpenBSD as `jemalloc` is not supported.

NOTE: I'm only 80% sure this works, as I was unable to finish compiling this on
my OpenBSD laptop: I got all the way to 603/617 and then ran out of memory (my
laptop is pretty ancient... like from 2012.) I'd love it if someone who has a
beefier OpenBSD box could give this a spin.

I'm also unsure if there are other parts of the repo that need to be updated to
ensure OpenBSD artifacts get built and distributed accordingly. Happy to make
changes to support that if folks could point me in the right direction.

closes #3672